scheduled-lambda/requirements.txt
aws-cdk-lib==2.155.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==2.28.1
cdk-lambda-powertools-python-layer==2.1.0